Change behavior of 'NoSpaces' settings flag: do not replace '_' with ' ' if flag is not set.
This was not a logical behavior: the description of the option is to
'replace spaces with underscores' if option is set; the inverse is to do
nothing and not to replace underscores with spaces.
If the replacement of underscores with spaces is needed (some comments
suggest this would be needed to cleanup some tags) it should be
implemented as a different option (and probably apply only to Tags).